Fix tool size

+++ project/src/main/java/pixel/tool/Tool.java
@@ -10,7 +10,7 @@ public interface Tool {
 	
 	public String getName();
 	
-	public void setSize(double size);
+	public void setSize(int size);
 	
 	public void use(Sprite sprite, MouseEvent event);
